Fix ExtractWebsiteContentBlock cost
